Moved standalone functions to math package
Improved documentation as i was at it.

Ref #9
