+++ src/neurom/math/__init__.py
@@ -1,3 +1,6 @@
 from neurom.math.inner import inner
 from neurom.math.integrate import integrate
 from neurom.math.jacobian import jacobian
+from neurom.math.trace import trace
+from neurom.math.identity import identity
+from neurom.math.transpose import transpose
